migration

+++ database/migrations/2018_10_05_101839_create_instruments_table.php
@@ -20,7 +20,7 @@ class CreateInstrumentsTable extends Migration
             $table->string('model');
             $table->integer('serial');
             $table->string('client');
-            $table->decimal('constant',5,2);
+            $table->double('constant',5,2);
             $table->date('calibrated');
             $table->integer('receipt');
             $table->timestamps();
